requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
chromadb
google-generativeai
sentence-transformers
python-dotenv